backend/requirements.txt
fastapi
uvicorn
python-docx
python-multipart
openai
markdownify
